fe-exersize/src/app/dashboard/dashboard.component.spec.ts
import { TestBed } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';
import { DashBoardComponent } from './dashboard.component';

describe('AppComponent', () => {
  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [
        RouterTestingModule
      ],
      declarations: [
        DashBoardComponent
      ],
    }).compileComponents();
  });

  it('should create the app', () => {
    const fixture = TestBed.createComponent(DashBoardComponent);
    const app = fixture.componentInstance;
    expect(app).toBeTruthy();
  });

  it(`should have as title 'fe-exersize'`, () => {
    const fixture = TestBed.createComponent(DashBoardComponent);
    const app = fixture.componentInstance;
    expect(app.title).toEqual('fe-exersize');
  });

  it('should render title', () => {
    const fixture = TestBed.createComponent(DashBoardComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('.content span')?.textContent).toContain('fe-exersize app is running!');
  });
});
